build: Fix linkage of libgcc
Some functions, like getpid() or kill(), weren't found during the
linking process because libgcc was linked after the lib group that was
defined in the Makefile. Groups allow libraries to be linked in any
order, and the linker will iterate all of them until all references have
been found.

This patch makes the linker stop adding libgcc implicitly, and it forces
the user to add it explicitly so that it can be part of the same group.