libsysnds: Implement basic FatFs cache
Right now this cache is super basic, but it is still useful because it
removes the hack to read from the SD of DSi that was present in
libsysnds. Now, all reads are done to main RAM because the cache is
allocated with malloc(), so there is never any read directed to DTCM,
where DMA can't write.