all: Initial refactored repository
Remove picolibc and avr-libstdcpp. Unfortunately, the history of
picolibc is very big. It isn't possible to prevent people from
downloading every time unless picolibc disappears from the history of
the repository.

For that reason, the history of the SDK has been condensed into one
commit. The following repository contains the old history of the SDK:

https://github.com/blocksds/sdk-pre-v0.4/tree/83e480fa8e702b1f97377e23ca026dd5ef41b91c

Also, the crts and default_arm7 submodules have been integrated in this
repository because their histories weren't very useful, and keeping them
as submodules made it harder to deal with all the changes done to them:

https://github.com/blocksds/crts/tree/41b8af2b4a03ea125901c07c87cd45a12e01da48

https://github.com/blocksds/default_arm7/tree/da0b1209b8384fb735a567d17b6da7635aca3767

The C++ example is broken at the moment because picolibc needs a patch
that hasn't been applied yet.