examples: tests: templates: Move keyboard example code
This example code shouldn't be in the template, it's unsafe and the
template isn't about showing how to use the keyboard, but about showing
how to integrate graphics, audio, etc.

Also, `scanf()` is a bad function to use for keyboard input. It blocks
the program and it doesn't check how big the buffer is.

For this reason, a new example has been introduced, that uses the
functions of libnds directly instead of libc functions.

A test has been added to make sure that the integration with the stdin
of libc is still working, as it may be useful in some simple
applications (think of a an application that wants to ask the user
where to save a file).