crts: Allocate more space for the scheduler stack
This is required so that it has enough space to handle interrupts
that may happen during the scheduler process. As interrupts may end up
being nested, the stack of the scheduler needs to have enough memory to
handle at least a few levels of nesting.

Normally, interrupts happen during threads other than the scheduler one,
so the stack used is the one of that thread. The default is 1 KB per
thread, which was enough to handle interrupts (but the scheduler only
had 256, so it wasn't enough).