templates: tests: libnds: Restore keyboard in template
Now that stdin handling is done correctly, it is possible to restore
this part of the template.

Also, all the warning text in the stdin test is no longer valid.
Because of that, it has been moved to the examples folder, as it is now
a reasonable way to get keyboard input by users.