docs: Update porting guide and introduce usage notes document
The text console now behaves in a similar way as the one in devkitARM,
so that section is outdated.

The part about placing objects in different sections or building code
as ARM instead of Thumb is no longer a difference, so it doesn't belong
in the porting guide. This "usage notes" document will eventually have
more tips to use BlocksDS, and may evolve to become an actual manual.