arm7: Reimplement ARM7 default code
The old code was unclear, undocumented, and it had copied some audio
initialization code that is already in libnds.

The code has been documented, and that audio initialization code is now
in libnds, not in the ARM7 main() function.

Also, this patch gets rid of the VCOUNT interrupt and it updates key
input in the VBLANK handler.