crts: Add script to preprocess the Teak crt0
The LLVM Teak toolchain comes with an assembler that can't evaluate
expressions. For example, if this is used:

    rst (1 << 7), mod0

The result is:

    ds_teak_crt0.s:44:5: error: Instruction could not be parsed
        rst (1 << 7), mod0

The user must write the resulting number instead. This makes the code
unreadable and unmaintainable. This script uses the GCC preprocessor to
expand macros, and a python script to evaluate expressions. That way the
original file can be easy to document, and it can be ready for the time
when the assembler can evaluate expressions.

Both the original and the preprocessed file will be kept in the
repository so that python isn't required to build the repository (to
reduce the size of the docker images).