submodules: Migrate to forks in BlocksDS organization
Instead of relying on repositories owned by their creators, fork them
so that BlocksDS doesn't depend on anything external to the BlocksDS
organization. The owner could change names, delete the repository...

It's still expected for users to contribute to the original repositories
instead of the forks. The forks are a last resort in case the owner
of the original repository isn't responsive, but normally they should
only track the original repositories without modifications.