build: submodules: Update architecture compiler flags
The new flags are the ones recommended for Wonderful Toolchain.

Also, ``-mtune=arm7tdmi`` is superfluous, so it has been removed.

Finally, add ``-Wl,--use-blx`` in the LDFLAGS of ARM9 binaries so that
the linker generates blx instructions instead of trampolines like for
the ARM7.

Update repositories to include the same changes.