templates: use build/$(NAME) for build directory
This allows flexibility in the event a user needs to build multiple
binaries from the same source code. For example, if one needs builds
with a specific DEFINE, and another without such a DEFINE, just a NAME
change will make sure these do not conflict, while also making it
relatively easy for a user to see which build directory is related to
which Makefile.

Test: compile