docs: examples: tests: Move documentation to Hugo
The reStructuredText documentation has been migrated to Markdown.

A script has been added to build the documentation. It builds the Doxygen
documentation of libnds, libteak, and BlocksDS and publishes it.

The links in examples and tests to documentation have been fixed. The links
in the documentation to examples and tests have been fixed as well.