examples: Replace models of 3D picking example
The original models didn't have a clear origin and, worse, they didn't come
with the original data used to generate them.

This commit adds Wavefront OBJ files with equivalent models (generated with
Blender, using the basic sample meshes you can add from the menu) and a
simple script to convert them to display lists (but it needs a script not
included in BlocksDS).