devkitarm_porting_guide: note relative paths outside PWD is not supported
For select "advanced" projects, such as a port of a source code that
was not designed with devkitARM in mind, it is common practice to have
a "platform" Makefile in a subdirectory that calls directories that
are out of the scope of said Makefile's PWD.

This behaviour is unsupported in BlocksDS at this time. It should be
noted.