examples: Fix VRAM bank setup in "text over 3D" example
libnds calls `vramDefault()` during the initialization of the library in
function `initSystem()` in file `source/arm9/system/initSystem.c`.

`vramDefault()` is located in `source/arm9/video/video.c`, and it sets up
VRAM A as background VRAM for the main engine.

The example was trying to setup VRAM H to value VRAM_F_MAIN_BG_0x06000000,
which is equivalent to VRAM_H_SUB_BG_0x06200000. This means that VRAM H
was being mapped as sub engine BG VRAM instead of main engine BG VRAM.
However, VRAM A was configured correctly, so the example worked.

The reason why the banks have to be unset is that when VRAM F is configured
correctly, it will cause a conflict with VRAM A, which needs to be
prevented. Normally VRAM A is used for textures so this conflict wouldn't
happen in a normal game, but this example doesn't use any texture, so the
resulting configuration of VRAM A is main BG VRAM.