examples: Stop using GL_RGB and TEXTURE_SIZE_xxx defines
GL_RGB shouldn't be used. It's just like GL_RGBA but glTexImage2D() is
forced to set bit 15 to 1 for all pixels, which is very slow. It's better
to expect the developer to convert the texture and set the bit 15
correctly.

TEXTURE_SIZE_xxx defines are very inconvenient to use, and now
glTexImage2D() accepts regular sizes in pixels, so it's better to use that.