examples: Add example of doing two-pass 3D
This can be used if the developer wants to duplicate the number of polygons
drawn on the screen at the same time.

The specific way to do two-pass 3D in this example (first draw the left
half of the screen, then the right one) is based on the ideas of Gericom.
The system normally documented online implies drawing the far polygons
first and overlying a new layer with the near polygons (but that means
losing Z-buffer support between the two passes).