examples: Cleanup of keyboard examples
The custom keyboard example now lets the user show and hide the keyboard
to show that the keyboard state is returned to the original one every
time it is shown.

The fields of the custom keyboard struct have been rearranged to match
the order in the documentation.

Some comments in the other examples were incorrect.