docker: Remove LLVM Teak toolchain from slim image
Most people don't need this toolchain, but every download of the Docker
image includes it. Removing it doesn't block people from installing it
in their own Dockerfile, but it saves time for everyone else. However,
this is a compatibility break for anyone using the slim image to build
DSP binaries.

This reduces the size of the slim image by 176 MB to 872 MB. The current
dev image is 1.36 GB in size.