examples: sys: templates: Fix Maxmod soundbank generation for NitroFS
When the makefiles were originally created, ndstool only accepted one
directory as the source of NitroFS files. A few months ago I added
support to provide multiple directories that are eventually merged into
one only NitroFS filesystem.

Originally, the only way to add the "soundbank.bin" file to NitroFS was
to add it to the nitrofs folder of the project of the user, which was a
terrible thing to do.

I never got around updating the Makefiles after adding support for
multiple directories to ndstool, so that hack stuck around.

This patch moves the "soundbank.bin" to its own folder in the "build"
folder and then it passes that folder and the user-provided NitroFS to
ndstool to combine their contents.

That way, the user folder will remain intact, and the clean rule can be
simplified.

Also, this allows us to simplify the gitignore files.