docker: Add ARM64 support to Docker image
This PR is to add support for linux/arm64 platforms on the Docker image.
This will allow users with 64 bit ARM computers (such as modern Macs and
Raspberry Pis) to use BlocksDS on Docker natively. Builds were 8x faster
on an M1 Mac mini, compared to emulation (16.526s vs 2m20.807s real
time).

One thing to note is that llvm-teak is not available on aarch64/ARM64,
so only the x86_64/AMD64 build includes the `toolchain-llvm-teak-llvm`
package.

The `build-docker.sh` script was also updated to do a multi-platform
build [1]. A single image is produced, that can be run for both
architectures. This process should be transparent to you on Docker
Desktop, emulation is handled automatically.

You can try the ARM64 builds on an AMD64 machine and vice-versa. For
example, `docker run --platform linux/arm64 -it --entrypoint bash
skylyrac/blocksds:dev-latest`. You may need to enable the contanierd
image store [2].

For testing I built the BlocksDS examples on both architectures.

[1] https://docs.docker.com/build/building/multi-platform/
[2] https://docs.docker.com/desktop/containerd/#build-multi-platform-images