docker: Install packages from repository
Currently BlocksDS is setup by installing the packages of the ARM compiler
and then building BlocksDS from source. This patch installs it from the
pacman packages instead to make the container easier to use. For example,
if you wanted to use µLibrary with the previous image, you'd have to
download the source code and build it. After this commit, you can just
install the right package.