docs: scripts: Document release process and update scripts
The `do-new-release-tasks.sh` script is now redundant, so it is being
removed as well. It will still be possible to build the images locally
with `build-docker.sh`, but there is no need for a script to do all the
new release tasks because now they are done on GitHub Actions.