examples: Build dynamic library example with make
The bash script previously used to build the example has been removed.

Note that this makefile isn't great, the DSL file and NDS ROM are always
built even if there have been no changes in the library. This happens
because the top level makefile can't keep track of the dependencies known
by makefiles in subfolders.