crts: Fix heap bounds in DSi mode
In DSi mode there is a lot of data at the end of RAM (check the memory map
in file docs/content/technical/memory_map.md). This is located right after
the end of DTCM because of how DTCM overlaps main RAM. Because of this,
fake_heap_end needs to be set to the start of DTCM, as that includes both
DTCM and the reserved memory.

In DS mode there is no overlap of DTCM with main RAM, so fake_heap_end
needs to be set to the end of main RAM minus the reserved memory.

The previous code included DTCM in the heap, but the stack is placed in
DTCM, so using too much heap would corrupt the stack.

Thanks @ds-sloth for the initial report that eventually found this bug.