examples: Fix example of streaming audio with Maxmod
Interrupt handlers are supposed to be quick. Sometimes it's unavoidable
that they take longer, but this is usually done for tasks that take a
long time to be processed, and the developer can enable nested
interrupts in that case if required.

Reading files from the filesystem is a very complicated process that can
involve communicating with the ARM7. This involves the FIFO helpers of
libnds, threading functions, loops that wait for interrupts, etc. This
is definitely too much for an interrupt handler.

For example, if the application is running on a DSi, the call stack may
look like this:

  Location | Functions
  ---------+-----------------------------------------------------------
  libnds   | IntrMain() ->
  Maxmod   | mmStreamUpdate() -> StreamExecuteUpdate() ->
  example  | streamingCallback() ->
  picolibc | fread("nitrofs:/...") ->
  libnds   | read() -> nitrofs_read() -> nitrofs_read_internal() ->
  picolibc | fread("sd:/...") ->
  libnds   | read() ->
  FatFs    | f_read() ->
  libnds   | disk_read() -> io->readSectors() / sdmmc_fifo_sectors() ->
  libnds   | fifoWaitValue32Async() -> swiIntrWait()

Doing all this work inside an interrupt handler will use a lot of stack,
and the problem only gets worse if nested interrupts are enabled.
`swiIntrWait()` enables interrupts internally, so it was very likely
that a new interrupt would be handled while waiting for the ARM7 to
finish reading data.

libnds has recently made a change to prevent `swiIntrWait()` from being
called if interrupts are disabled [1], but this only fixes the issue of
nesting interrupts, it doesn't solve the problem of the complexity.

The best thing to avoid this issue is to move the complexity to the main
loop of the example and use a circular buffer to pass data to the Maxmod
callback.

[1] https://github.com/blocksds/libnds/commit/a0c2f359184c86661b928a25303954beb7c01a52