crt0: Flush cpu caches after all the bss and arm9i binary are set up
The arm9 crt0 writes the arm9i binary to ram after the MPU has been set up, thus
writing to the cache.
Without the explicit flush, it could happen that the user code calls functions from,
or access data in this area that hasn't been flushed to ram, leading to crashes
and erroneous behavior.