version: Add more makefile helpers
This makes it so blocksds_version_equals, blocksds_version_at_least and blocksds_version_at_most return 0 or 1 when called.
This is because previously the makefile would just error out if you used $(eval $(call ...)), with no way to do anything to handle the case for the user.
Now, the user can $(call ...) the functions to decide what to do with their result.

Then, it introduces the helper functions error_if_blocksds_version_different, error_if_blocksds_version_less_than and error_if_blocksds_version_greater_than, which an user may call to use the default error handling.

This improves flexibility, due to separating the logic from the error handling.

It also improves the readme by adding to the makefile example proper handling of the older release versions.