crts: Improve synchronization before calling ARM7 main()
Currently, right before jumping to the ARM7 main() the ARM7 uses IPC_SYNC
to send a zero to the ARM9. The ARM9 then reads that value and it
continues booting.

However, if the ARM7 user code writes to the IPC_SYNC register right at
the beginning of main() the ARM9 will never see the 0 and it will hang.

This patch modifies the code so that the ARM7 also needs to wait for the
ARM9 to send a zero. The ARM9 doesn't use the IPC_SYNC register after that,
so the ARM7 can safely jump to main() as soon as it sees the zero.

The ARM9 still takes quite a while to reach main() after this handshake,
so there are no potential issues there.