default_makefiles: templates: Remove hardcoded C/C++ standard version
Let the user decide the standard version they want to use. If the user
doesn't specify one, let GCC decide. This is rarely a problem in C
projects, but it can be a bit more problematic in C++ projects.

Users that want to keep using old standards will need to manually set it
to the right standard in their makefiles.

The Teak toolchain is a bit older and it will probably stay like that
for the time being, so no change has been done to the Teak makefiles.